Refactor: Move semantic analyzer pass 1 to a new module
Just move things around, tweak inter-module references and add
necessary imports.

The first pass still directly accesses the second pass so there is
more work to do.
