+++ mypy_self_check.ini
@@ -12,5 +12,8 @@ warn_unused_ignores = True
 [mypy-mypy.semanal]
 strict_optional = False
 
+[mypy-mypy.semanal_pass1]
+strict_optional = False
+
 [mypy-mypy.semanal_pass3]
 strict_optional = False
